İlkMetot

+++ CalorieCalculate/Model/Entities/UserInformation.cs
@@ -12,8 +12,8 @@ namespace CalorieCalculate.Model.Entities
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public DateTime BirthDate { get; set; }
-        public int Height { get; set; }
-        public int Weight { get; set; }
+        public decimal Height { get; set; }
+        public decimal Weight { get; set; }
 
         #region Navigation
